requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
crewai==0.1.0
pydantic==2.5.0
sqlalchemy==2.0.23
cryptography==41.0.7
PyJWT==2.8.1
python-dotenv==1.0.0
requests==2.31.0
